refactor(codegen): move solidity functions to solidity.rs
